StreamlitLLMs/TextSummurizerStreamlitLLM/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TextSummurizerStreamlitLLM"
version = "0.0.1"
description = "A StreamLit App to summurize a long text loaded from a text file"
authors = [{name = "Mansy", email = "mansy.soroush@gmail.com"}]
dependencies = [
    "ipykernel",
    "python = 3.11.4",
    "python-dotenv",
    "streamlit >= 1.37.0",
    "langchain >= 0.2.11",
    "langchain-openai >= 0.1.19"
]